span: Resolve conversion warning
[Why]
If compiled with clang and -Wconversion warnings enable, the
compiler will complain:
> gul14/span.h:218:42: warning: implicit conversion changes signedness: 'long' to 'std::size_t' (aka 'unsigned long') [-Wsign-conversion]
>   218 |         : storage_(first_elem, last_elem - first_elem)
>       |           ~~~~~~~~             ~~~~~~~~~~^~~~~~~~~~~~

+++ include/gul14/span.h
@@ -215,7 +215,7 @@ public:
 
     /// See [std::span](https://en.cppreference.com/w/cpp/container/span).
     constexpr span(pointer first_elem, pointer last_elem)
-        : storage_(first_elem, last_elem - first_elem)
+        : storage_(first_elem, static_cast<std::size_t>(last_elem - first_elem))
     {}
 
     /// See [std::span](https://en.cppreference.com/w/cpp/container/span).
